+++ memtable.go
@@ -10,11 +10,11 @@ func newMemtable() *memtable {
 	}
 }
 
-func (m *memtable) Put(key Key, value []byte, seqnum uint32) {
+func (m *memtable) Put(key Key, value []byte, seqnum uint64) {
 	m.list.Put(key, value, seqnum)
 }
 
-func (m *memtable) Delete(key Key, seqnum uint32) {
+func (m *memtable) Delete(key Key, seqnum uint64) {
 	m.list.Delete(key, seqnum)
 }
 
@@ -22,7 +22,7 @@ func (m *memtable) Len() int {
 	return m.list.Len()
 }
 
-func (m *memtable) Get(key Key, snapshotSeq uint32) ([]byte, error) {
+func (m *memtable) Get(key Key, snapshotSeq uint64) ([]byte, error) {
 	return m.list.Get(key, snapshotSeq)
 }
 
@@ -30,7 +30,7 @@ func (m *memtable) ByteSize() uint64 {
 	return m.list.ByteSize()
 }
 
-func (m *memtable) Scan(lo, hi Key, snapshotSeq uint32) *skiplistIterator {
+func (m *memtable) Scan(lo, hi Key, snapshotSeq uint64) *skiplistIterator {
 	return m.list.Scan(lo, hi, snapshotSeq)
 }
